refactor(4d): P0.2 — extract playback state into its own slice
Playback concerns (animation toggle, play state, cursor time, speed,
loop, animationSettings + 11 related mutators) split out of
scheduleSlice.ts into a new playbackSlice.ts (~130 lines).

scheduleSlice.ts drops ~70 lines — now focused on schedule-domain
state + mutators. Playback mutators no longer contest for attention
with the edit-pipeline comments.

No consumer changes: every component still reads via
`useViewerStore(s => s.playbackTime)` — the combined-store shape
includes both slices, so imports stay identical.

The rAF advance loop still needs `scheduleRange` for its end-of-
schedule clamp; the playback slice declares that as a cross-slice
read the same way scheduleSlice declares `dirtyModels` /
`mutationVersion` reads (no `as unknown as` casts).

Tests: 487 viewer green. Build unchanged.
